requirements.txt
aiogram==3.13.1
sqlalchemy==2.0.35
asyncpg==0.30.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
requests==2.32.3
